Pearl: fix shortcut-loop detection on squares.
The final deduction technique in Tricky mode is to identify a
particular square and rule out one of the possible states it could be
in, by observing that that state would connect together two
already-connected things - hence forming a loop - but the loop would
be too small to include all the known non-blank squares, and therefore
couldn't be the true solution.

There were two bugs in this code. One would have made it rule out many
perfectly sensible square states and conclude that the puzzle had no
solution, if it hadn't been for the other bug, which prevented the
code from ever doing anything!

The 'never do anything' bug: in the small sub-loop that looked for the
common equivalence class of the two neighbouring squares we'd be
connecting, the output variable 'e' never had a valid value assigned
to it, because the assignment which should have done that was written
backwards, 'ee = e' instead of 'e = ee'.

With that fixed, the 'rule out perfectly sensible things' bug became
visible: if the square is already _known_ to connect a particular pair
of neighbours together, then it's not interesting that they're in the
same equivalence class, and doesn't mean we're about to form a loop at
all!

With both bugs fixed, a test run of 100 puzzle generations in Tricky
mode, 'pearl --generate 100 8x8dt#12345', generates 18 puzzles
differently after the change than before. All of those 18 puzzles
require this now-enabled deduction, and the old Pearl can't solve them
at all.