X11: use the right timestamp for SetSelectionOwner.
When you set an X selection in response to an X event, you're supposed
to take the X server timestamp from the event and pass it back to the
SetSelectionOwner request, instead of just passing the cop-out value
CurrentTime.

I noticed the use of CurrentTime in a line modified by the previous
commit, and thought I'd see if it was easy to fix. It turns out it is:
GTK has a handy gtk_get_current_event_time() function, which will
return the event timestamp if you call it from within an event handler
responding to an X event, and default to CurrentTime if one isn't. So
it's very easy to tell GTK "please do the right thing if convenient".

(I presume that when GTK is using a backend other than X11, either all
of this is ignored anyway, or the same function call will find
something appropriate to do.)