Mines: allow params to control first-click location.
The string encoding of game_params can now include a location like
"X0Y5" at the end. This has no effect on interactive play (where the
game waits to see where the player first clicks), but in command-line
'mines --generate', it forces the pre-played first click to be at the
specified coordinates instead of randomised.

The idea is that this makes it easier to reproduce bugs that came up
in interactive play. For example, I ran across a bug this morning
which seems to come up more often when the first click is near a
corner.