Mines: fix a hang during grid generation.
The hang is reproducible, as of the code before this commit, by
running 'mines --generate 1 9x15n55X1Y1#2-94'.

What seems to happen is that the combination of [solver with
perturbations mixed in] gets stuck in a loop. At one end of the loop
there's a large square area cleared around the initial click, with
every square on the boundary being a mine. So the player can certainly
solve up to the boundary - but beyond the boundary they can't tell
what's a mine and what's not. So the perturber thinks about what to
do, and decides to fill _everything_ in the already-solved clear area
with mines from the unsolved area (leaving only a 3×3 region around
the initial click). That's no better, so the next run of the perturber
clears one extra row and column, and the next one does the same, and
so on, until we're back in the first situation. And since every time
we go round this loop we re-mark some squares as "not solved yet" and
then solve them, it looks as if progress is being made, and we don't
abandon the entire effort and restart from scratch, which is the usual
solution to any "stuck in a local optimum" problem.

This commit applies a simple and hopefully robust fix to get out of
the rut: we introduce a new notion of "progress" which doesn't count
re-solving a square that was solved before and then marked as undone
because of a perturbation. Now, after a couple of iterations around
that loop, the new nperturbs_since_last_new_open mechanism will notice
that no square has been opened _that was never opened before_ in too
long, and pull the plug.

I'm not at all happy with this fix. It should solve the immediate
hang, but I think probably what I _ought_ to do is rethink the entire
strategy of perturbation, in the large-scale case where we've run out
of perturbations we can make within a single 3×3 block and have to
fall back to either emptying or filling every single unknown square.

But fixing the immediate problem is a start!