16 KB page size compatibility: use newer NDK
Otherwise Studio picks NDK 21 which will not be 16 KB aligned.

"Starting November 1st, 2025, all new apps and updates to existing apps
submitted to Google Play and targeting Android 15+ devices must support
16 KB page sizes on 64-bit devices."

https://developer.android.com/guide/practices/page-sizes

Recording the NDK version used to build each release is also an
improvement in the reproducibility of the build.