slant: make fill_square() fallible.
Donald Hobson reported that the game id 2x2:020c0a0, which has no
legal solution, would be 'solved' uncomplainingly, by returning a
solution containing a loop. This happened because the final edge of
the loop was filled in from the vertex clues, without checking to see
if it _also_ violated the loop constraint.

Now fill_square() checks whether that's happened, and returns a
boolean to indicate failure. Every call site checks that return value,
and declares the puzzle inconsistent if it failed. So this kind of
situation, where neither state of a square is valid but for completely
different reasons, should now be detected.

(See also commit f21d3e4c74b2338, which fixed a similar kind of bug in
Group: if the final grid square was filled in by a general latin.c
deduction, but introduced a violation of the Group-specific
constraints at the last moment, that violation wouldn't be detected
and the non-solution would be considered acceptable. In that case it
could happen during backtracking search, leading to a puzzle being
reported as ambiguous.)

While I'm here, I also double-check the case where fill_square is
called a second time on an already filled square. It still returns
success if the square is being filled with the same contents as
before, but now it returns failure if the caller is trying to
contradict the square's existing contents. I have no example in which
this happens, but it seemed like a worthwhile precaution while I was
there anyway.