Filling: make Backspace clear the selected grid squares.
A user pointed out that it currently has the same behaviour as Esc: it
just _deselects_ the selected squares, without changing their
contents. That seems inconsistent with its use in most of the other
'click to select a square and then type something to put in it' user
interfaces in this collection, like Solo, Keen, Towers, ... in which
generally selecting a square and pressing Backspace clears whatever
the user had previously written in that square.

This behaviour was introduced in commit c469bd285cc4f48, ten years
ago, at the same time as introducing the general keyboard-only
control. It seems to have been semi-deliberate: the patch documented
the new behaviour, but didn't notice that it contradicted the existing
behaviour (also documented, in the previous paragraph) of Backspace
clearing a single cell you'd just clicked in.