Palisade: fix error reports in clear_complete_regions mode.
The user preference introduced in commit 7fa03051562ee81 causes any
'uncertain' edge to be shown as 'off' as soon as the user has filled
in the entire border of a correctly sized region with that edge in the
interior. This was taken into account for display purposes, but not
for purposes of highlighting errors in the clues.

So if you play Palisade without clicking edges into the 'off' state,
relying on that preference to clear them for you, and you 'solve' a
grid in an incorrect way that leaves a clue with too _few_
neighbouring region edges, then the clue wouldn't light up red as an
error. Turning off clear_complete_regions would reveal the cause:
there are still edges around the square marked 'uncertain'.

Now the edges shown as 'off' in the display are also treated as 'off'
for the purposes of clue error checking, so as soon as you complete a
region with an undersatisfied clue, the clue lights up.