New specialist backend function, set_public_desc().
This is the counterpart of midend_supersede_game_desc(): in the rare
case where a puzzle needs to set separate private and public game
descriptions, this method is used during game reload to allow the
contents of the _public_ description to affect the starting game
state.

I hope this will be even less often used than the rest of the
Mines-specific bodges. But Mines itself _is_ about to use it, in the
next commit.