Mines: mark the first click when undoing past it.
One hole in Mines's "you never have to guess" story can occur as a UI
accident, rare but possible: if a user clicks on a blank grid causing
a game to be generated, and then (perhaps absentmindedly) presses Undo
to return to a blank grid, then the mine locations are now fixed, and
clicking somewhere _else_ might kill you, or open a single square,
or (perhaps even more confusing) open a region from which the game
can't be solved.

This commit does not stop the user doing that, but makes it clear when
they did. Now, after the first-click location has been selected and
the mines have been placed, if you undo the first move, the location
of your original first click will be marked with an X. Then if you
click somewhere else and find something confusing has happened, it
will at least be more obvious what's going on.